adding authenticated routes with jwt

+++ backend/src/db/services/posts.js
@@ -1,11 +1,10 @@
-import { Post } from '../models/post.js'
+import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
-
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
-
 
 async function listPosts(
   query = {},
@@ -18,8 +17,10 @@ export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
 
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
 
 export async function listPostsByTag(tags, options) {
@@ -30,14 +31,18 @@ export async function getPostById(postId) {
   return await Post.findById(postId)
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(
+  userId,
+  postId,
+  { title, contents, tags },
+) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
